requirements.txt
# Instagram Auto-Reply Service - Python Requirements
# ==================================================
# Install with: pip install -r requirements.txt
# Python Version: >= 3.9

# Core Flask Framework
flask>=3.0.0,<4.0.0
werkzeug>=3.0.0,<4.0.0

# Database
flask-sqlalchemy>=3.1.0,<4.0.0
sqlalchemy>=2.0.0,<3.0.0

# HTTP Client
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0

# Security - Required for token encryption
cryptography>=42.0.0

# Environment Configuration
python-dotenv>=1.0.0,<2.0.0

# Production WSGI Server
gunicorn>=21.0.0,<23.0.0
